add confirmation modal for delete

+++ src/app/heroes/hero-list/hero-list.component.ts
@@ -1,23 +1,25 @@
-import { Component, computed, signal } from '@angular/core';
+import { Component, computed, inject, signal } from '@angular/core';
 import { HeroService } from '../../services/hero.service';
-import { CommonModule, NgFor, NgIf } from '@angular/common';
+import { CommonModule} from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { PaginationComponent } from '../../shared/components/pagination/pagination.component';
+import { ModalService } from '../../services/modal.service';
+import { ConfirmationModalComponent } from '../../shared/components/confirmation-modal/confirmation-modal.component';
 
 @Component({
   standalone: true,
   selector: 'app-hero-list',
-  imports: [CommonModule, RouterModule, NgFor, NgIf, PaginationComponent],
+  imports: [CommonModule, RouterModule, ConfirmationModalComponent, PaginationComponent],
   templateUrl: './hero-list.component.html',
   styleUrls: ['./hero-list.component.scss'],
 })
 export class HeroListComponent {
-  heroes;
-  filterText = signal<string>('')
+  private heroService = inject(HeroService)
+  private modalService = inject(ModalService);
 
-  constructor(private heroService: HeroService) {
-    this.heroes = this.heroService.heroes;
-  }
+  heroes = this.heroService.heroes
+
+  filterText = signal<string>('')
 
   filteredHeroes = computed(() =>
     this.heroes().filter((hero) =>
@@ -33,8 +35,11 @@ export class HeroListComponent {
     return this.heroService.totalPages();
   }
 
-  deleteHero(id: number) {
-    this.heroService.deleteHero(id);
+  async deleteHero(id: number) {
+    const confirmed = await this.modalService.open();
+    if (confirmed) {
+      this.heroService.deleteHero(id);
+    }
   }
 
   onFilterChange(event: Event) {
